refactor(preview-actions): extract refresh prefetch and line-count flow
